Fixed warnings and errors in the JS code and updated JS bundle.
