--- CHEWBBACA/SchemaEvaluator/resources/loci_reports/src/pages/ReportPage.css
@@ -1,8 +1,0 @@
-:local(.mainDiv) {
-	margin-top: 10px;
-	margin-bottom: 10px;
-}
-
-:local(.secondaryDiv) {
-	margin-top: 20px;
-}
