+++ CHEWBBACA/report_template_components/src/components/AlertMUI.js
@@ -2,10 +2,10 @@ import Alert from '@mui/material/Alert';
 import Typography from '@mui/material/Typography';
 
 
-const AlertMUI = ({ severity, fontSize, children }) => {
+const AlertMUI = ({ ind, severity, fontSize, children }) => {
 	return (
-		<Alert severity={severity}>
-			<Typography sx={{ fontSize: fontSize }}>
+		<Alert key={ind} severity={severity}>
+			<Typography key={ind} sx={{ fontSize: fontSize }}>
 				{children}
 			</Typography>
 		</Alert>
