+++ CHEWBBACA/report_template_components/src/components/PlotlyPlot.js
@@ -5,13 +5,13 @@ import Plot from "react-plotly.js";
 const PlotlyPlot = ({ plotData, layoutProps, configOptions, ...rest }) => {
 	return (
 		<Plot
-        data={plotData}
-        layout={layoutProps}
-        config={configOptions}
-        useResizeHandler={true}
-        style={{ width: "100%", height: "100%" }}
-        {...rest}
-    />
+            data={plotData}
+            layout={layoutProps}
+            config={configOptions}
+            useResizeHandler={true}
+            style={{ width: "100%", height: "100%" }}
+            {...rest}
+        />
 	)
 };
