+++ CHEWBBACA/report_template_components/src/components/TabPanelMUI.js
@@ -59,7 +59,6 @@ const TabPanelMUI = ({ ContentTitles, ContentData }) => {
 		)
 	});
 
-
 	return (
 		<Box sx={{ width: "100%" }}>
 			<Paper elevation={3}>
