+++ CHEWBBACA/report_template_components/src/pages/SchemaEvaluator/schema_report/ReportPage.css
@@ -1,0 +1,8 @@
+:local(.mainDiv) {
+	margin-top: 10px;
+	margin-bottom: 10px;
+}
+
+:local(.secondaryDiv) {
+	margin-top: 20px;
+}
